Add a new "Chef Metadata" language and snippets
I created a new language and snippets for Chef Metadata. Why a lang instead of just adding them to the existing Chef language? Well this way we accurately identify metadata files and we only add these snippets there. It's a UX that I'd like to continue with custom resources in the future. I also created everything in JSON since that seems to be the vscode standard. I can convert the plist file for the existing chef stuff sometime in the future.

Signed-off-by: Tim Smith <tsmith@chef.io>